use 32 bit types in event queue data, and use queue item data better

+++ include/openmc/cuda/advance.h
@@ -21,7 +21,7 @@ __global__ void process_advance_events_device(EventQueueItem* queue,
   unsigned tid = threadIdx.x + blockDim.x * blockIdx.x;
   bool surface = false;
   bool collision = false;
-  int64_t p_idx;
+  unsigned p_idx;
   Particle* p;
 
   if (tid < queue_size) {
